requirements.txt
jinja2
python-docx
markdown
numpy
scikit-learn
flask
flask-cors
pytest
pytest-flask
pytest-cov
pytest-html
openai
anthropic>=0.25.0
python-dotenv
tiktoken>=0.5.0
